refactor CML and add control folder in components (#190)
* refactor CML and add control folder in components

* Add index.js in Controls

* Moved leaflet utils to a separate utils-directory
