+++ src/lib/components/NewLayeredMap/components/controls/components/VerticalZoom/L.VerticalZoom.js
