+++ src/lib/components/NewLayeredMap/components/controls/index.js
@@ -1,0 +1,1 @@
+export { default } from './Controls';
