+++ src/lib/components/NewLayeredMap/utils/index.js
@@ -1,7 +1,9 @@
 import * as ColorScaleUtils from './colorScale';
 import * as ImageUtils from './image';
+import * as LeafletUtils from './leaflet';
 
 export default {
     ...ColorScaleUtils,
     ...ImageUtils,
+    ...LeafletUtils,
 }
